FolderCard: Restructure component files
