+++ packages/core/upload/admin/src/components/FolderCard/FolderCard/index.js
@@ -1,0 +1,1 @@
+export * from './FolderCard';
