+++ packages/core/upload/admin/src/components/FolderCard/FolderCardBody/FolderCardBody.js
@@ -3,7 +3,7 @@ import styled from 'styled-components';
 
 import { Flex } from '@strapi/design-system/Flex';
 
-import { useFolderCard } from './FolderCardContext';
+import { useFolderCard } from '../contexts/FolderCard';
 
 const StyledBox = styled(Flex)`
   user-select: none;
