+++ packages/core/upload/admin/src/components/FolderCard/FolderCardBody/index.js
@@ -1,0 +1,1 @@
+export * from './FolderCardBody';
