+++ packages/core/upload/admin/src/components/FolderCard/FolderCardCheckbox/index.js
@@ -1,0 +1,1 @@
+export * from './FolderCardCheckbox';
