+++ packages/core/upload/admin/src/components/FolderCard/contexts/FolderCard.js
@@ -1,6 +1,6 @@
 import { createContext, useContext } from 'react';
 
-export const FolderCardContext = createContext();
+export const FolderCardContext = createContext({});
 
 export function useFolderCard() {
   return useContext(FolderCardContext);
