+++ packages/core/upload/admin/src/components/FolderCard/hooks/useId.js
@@ -4,8 +4,8 @@ let id = 0;
 
 const genId = () => ++id;
 
-const useId = (prefix, initialId) => {
-  const idRef = useRef(initialId || `${prefix}-${genId()}`);
+const useId = initialId => {
+  const idRef = useRef(`${initialId}-${genId()}`);
 
   return idRef.current;
 };
