+++ packages/core/upload/admin/src/components/FolderCard/index.js
@@ -1,4 +1,5 @@
 export { FolderCard } from './FolderCard';
 export { FolderCardBody } from './FolderCardBody';
 export { FolderCardCheckbox } from './FolderCardCheckbox';
-export { FolderCardLink } from './FolderCardLink';
+
+export { FolderCardLink } from './styled';
